+++ frontend/app/src/components/helper/AssetDetails.vue
@@ -1,25 +1,35 @@
 <script setup lang="ts">
 import { type ComputedRef } from 'vue';
 import { type AssetInfoWithId } from '@/types/asset';
+import { getAddressFromEvmIdentifier } from '@/utils/assets';
 
-const props = defineProps({
-  asset: {
-    required: true,
-    type: String,
-    validator: (value: string): boolean => !!value && value.length > 0
-  },
-  assetStyled: { required: false, type: Object, default: () => null },
-  opensDetails: { required: false, type: Boolean, default: false },
-  hideName: { required: false, type: Boolean, default: false },
-  dense: { required: false, type: Boolean, default: false },
-  enableAssociation: { required: false, type: Boolean, default: true },
-  isCollectionParent: { required: false, type: Boolean, default: false }
-});
+const props = withDefaults(
+  defineProps<{
+    asset: string;
+    assetStyled?: Record<string, unknown>;
+    opensDetails?: boolean;
+    hideName?: boolean;
+    dense?: boolean;
+    enableAssociation?: boolean;
+    isCollectionParent?: boolean;
+    link?: boolean;
+  }>(),
+  {
+    assetStyled: undefined,
+    opensDetails: false,
+    hideName: false,
+    dense: false,
+    enableAssociation: true,
+    isCollectionParent: false,
+    link: false
+  }
+);
 
 const { asset, enableAssociation, isCollectionParent } = toRefs(props);
 const { assetInfo } = useAssetInfoRetrieval();
 
 const assetDetails = assetInfo(asset, enableAssociation, isCollectionParent);
+const address = reactify(getAddressFromEvmIdentifier)(asset);
 
 const currentAsset: ComputedRef<AssetInfoWithId> = computed(() => ({
   ...get(assetDetails),
@@ -28,7 +38,7 @@ const currentAsset: ComputedRef<AssetInfoWithId> = computed(() => ({
 </script>
 
 <template>
-  <div>
+  <div class="flex-row d-flex">
     <asset-details-base
       :hide-name="hideName"
       :asset="currentAsset"
@@ -38,6 +48,13 @@ const currentAsset: ComputedRef<AssetInfoWithId> = computed(() => ({
       :enable-association="enableAssociation"
       :show-chain="!isCollectionParent"
       :is-collection-parent="isCollectionParent"
+    />
+    <hash-link
+      v-if="link && address"
+      type="address"
+      :evm-chain="assetDetails?.evmChain"
+      link-only
+      :text="address"
     />
   </div>
 </template>